Endpoint logic fixed

+++ app/student_reviews/dataclasses.py
@@ -6,9 +6,9 @@ from typing import Optional
 class StudentReview(BaseModel):
     id: int
     created_at: datetime
-    tutor_id: str
-    student_id: str
-    rating: float
+    tutor_id: str  # UUID stored as string
+    student_id: str  # UUID stored as string
+    rating: int  # Changed from float to int to match the database
     comment: Optional[str] = None
 
     class Config:
